Add support for Plugs, improve structure
